flask/requirements.txt
Flask==2.1.0
Flask-Cors==3.0.9
Flask-JSON==0.3.4
html5lib==1.1
httplib2==0.18.1
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
uritemplate==3.0.1
urllib3==1.26.2
Werkzeug==2.1.2
gunicorn==20.1.0
uWSGI==2.0.20
python-dotenv==0.21.0
lxml==4.6.3
numpy==1.21.3
